backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
python-multipart==0.0.9
pydantic==2.4.2
email-validator==2.1.0.post1
pydantic-settings==2.0.3
psycopg[binary,pool]==3.1.12
redis[hiredis]==5.0.1
structlog==23.1.0
httpx==0.24.1
aio-pika==9.2.2
prometheus-fastapi-instrumentator==6.1.0
celery==5.3.4
minio==7.1.16
Pillow==10.0.1
sentry-sdk==1.32.0
python-dotenv==1.0.0
